--- a/project/annotations/urls.py
+++ b/project/annotations/urls.py
@@ -21,6 +21,8 @@ image_urlpatterns = [
 ]
 
 source_urlpatterns = [
+    path('export/',
+         views.AnnotationsExportView.as_view(), name="export_annotations"),
     path('batch_delete_ajax/',
          views.batch_delete_annotations_ajax,
          name="batch_delete_annotations_ajax"),