--- a/project/export/forms.py
+++ b/project/export/forms.py
@@ -1,18 +1,6 @@
 from django.forms import Form
-from django.forms.fields import ChoiceField, MultipleChoiceField
-from django.forms.widgets import CheckboxSelectMultiple, RadioSelect
-
-
-class ExportAnnotationsForm(Form):
-    optional_columns_choices = (
-        ('annotator_info', "Annotator info"),
-        ('machine_suggestions', "Machine suggestions"),
-        ('metadata_date_aux', "Image metadata - date and auxiliary fields"),
-        ('metadata_other', "Image metadata - other fields"),
-    )
-    optional_columns = MultipleChoiceField(
-        widget=CheckboxSelectMultiple, choices=optional_columns_choices,
-        required=False)
+from django.forms.fields import ChoiceField
+from django.forms.widgets import RadioSelect
 
 
 class ExportImageStatsForm(Form):