--- a/project/export/urls.py
+++ b/project/export/urls.py
@@ -4,8 +4,6 @@ from . import views
 urlpatterns = [
     path('metadata/',
          views.export_metadata, name="export_metadata"),
-    path('annotations/',
-         views.export_annotations, name="export_annotations"),
     path('image_covers/',
          views.ImageCoversExportView.as_view(), name="export_image_covers"),
     path('labelset/',