… project/annotations/tests/test_delete.py
… project/annotations/tests/test_export_csv.py
… project/calcification/tests/test_export_stats.py
… project/cpce/tests/test_export.py
… project/cpce/utils.py
… project/export/templates/export/annotation_export_wait_times.html
… project/export/tests/test_covers.py
… project/visualization/static/js/tests/BrowseImagesActionsTest.js
… project/visualization/templates/visualization/browse_images.html
… project/visualization/templates/visualization/browse_images_actions.html
… project/visualization/tests/test_browse_images.py
… project/visualization/tests/test_delete_images.py
… project/visualization/tests/utils.py